#1248 - extracted interface in dto, replaced ternary condition with usage of isAllDatesAreSelected() method, some code cleanup
