+++ sandbox/projects/h21-be-ui-kit/src/dto/i-dialog-data.ts
@@ -1,0 +1,11 @@
+export interface DialogData {
+	rangeSelectMode: boolean;
+	fromDateText: string;
+	toDateText: string;
+	startDate: Date;
+	finishDate: Date;
+	fromDate: Date;
+	toDate: Date;
+	selectedFromDate: Date;
+	selectedToDate: Date;
+}
